pyproject.toml
[tool.poetry]
name = "createmeeting"
version = "0.1.0"
description = ""
authors = ["likerk <nagorniykirill@gmail.com>"]
packages = [
	{ include = "bot" },
]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "bot.scripts.main:main"
